data-processor/requirements.txt
pymongo==3.11.0
Pympler==0.9
python-rake==1.5.0
nltk==3.5